+++ src/main/java/com/github/theredbrain/bettercombatextension/mixin/client/network/ClientPlayerEntityMixin_BetterCombatReplacementMixin.java
@@ -1,11 +1,14 @@
 package com.github.theredbrain.bettercombatextension.mixin.client.network;
 
 import com.github.theredbrain.bettercombatextension.BetterCombatExtension;
+import com.github.theredbrain.bettercombatextension.bettercombat.DuckWeaponAttributesAttackMixin;
 import com.github.theredbrain.bettercombatextension.client.DuckMinecraftClientMixin;
 import com.mojang.authlib.GameProfile;
 import net.bettercombat.BetterCombatMod;
+import net.bettercombat.api.AttackHand;
 import net.bettercombat.api.MinecraftClient_BetterCombat;
 import net.bettercombat.config.ServerConfig;
+import net.bettercombat.logic.PlayerAttackHelper;
 import net.bettercombat.utils.MathHelper;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
@@ -56,14 +59,37 @@ public abstract class ClientPlayerEntityMixin_BetterCombatReplacementMixin exten
 	public void bettercombatextension$tickMovement(CallbackInfo ci) {
 		boolean isWeaponSwingInProgress = ((MinecraftClient_BetterCombat) this.client).isWeaponSwingInProgress();
 		ServerConfig config = BetterCombatMod.config;
-		double multiplier = Math.min(Math.max((double) config.movement_speed_while_attacking, 0.0), 1.0);
+		com.github.theredbrain.bettercombatextension.config.ServerConfig betterCombatExtensionServerConfig = BetterCombatExtension.SERVER_CONFIG;
+		double multiplier = Math.min(
+				Math.max(
+						(double) config.movement_speed_while_attacking,
+						betterCombatExtensionServerConfig.minimum_global_attack_movement_speed_multiplier.get()
+				),
+				betterCombatExtensionServerConfig.maximum_global_attack_movement_speed_multiplier.get()
+		);
 		ItemStack activeItemStack = this.getStackInHand(((DuckMinecraftClientMixin) this.client).bettercombatextension$getCurrentAttackHand());
 		boolean isMovementPenaltyIgnored = activeItemStack.isIn(BetterCombatExtension.IGNORES_ATTACK_MOVEMENT_PENALTY) && isWeaponSwingInProgress;
 		ClientPlayerEntity clientPlayer = (ClientPlayerEntity) (Object) this;
-		if (multiplier != 1.0 && !isMovementPenaltyIgnored) {
+		MinecraftClient_BetterCombat client = (MinecraftClient_BetterCombat) MinecraftClient.getInstance();
+
+		// attack specific movement modifier
+		double attack_specific_modifier = 1.0;
+		AttackHand attackHand = PlayerAttackHelper.getCurrentAttack(clientPlayer, client.getComboCount());
+		if (attackHand != null) {
+				attack_specific_modifier = Math.min(
+				Math.max(
+						((DuckWeaponAttributesAttackMixin) (Object) attackHand.attack()).bettercombatextension$getMovementSpeedMultiplier(),
+						betterCombatExtensionServerConfig.minimum_attack_specific_movement_speed_multiplier.get()
+				),
+				betterCombatExtensionServerConfig.maximum_attack_specific_movement_speed_multiplier.get()
+		);
+		}
+
+		if ((attack_specific_modifier != 1.0 || multiplier != 1.0) && !isMovementPenaltyIgnored) {
 			if (!clientPlayer.hasVehicle() || config.movement_speed_effected_while_mounting) {
-				MinecraftClient_BetterCombat client = (MinecraftClient_BetterCombat) MinecraftClient.getInstance();
+//				MinecraftClient_BetterCombat client = (MinecraftClient_BetterCombat) MinecraftClient.getInstance();
 				float swingProgress = client.getSwingProgress();
+
 				if ((double) swingProgress < 1/*0.98*/) {
 					if (config.movement_speed_applied_smoothly) {
 						double p2 = 0.0;
@@ -73,17 +99,21 @@ public abstract class ClientPlayerEntityMixin_BetterCombatReplacementMixin exten
 							p2 = MathHelper.easeOutCubic(1.0 - ((double) swingProgress - 0.5) * 2.0);
 						}
 
-						multiplier = (double) ((float) (1.0 - (1.0 - multiplier) * p2));
+						if (multiplier != 1.0) {
+							multiplier = (double) ((float) (1.0 - (1.0 - multiplier) * p2));
+						}
+						if (attack_specific_modifier != 1.0) {
+							attack_specific_modifier = (double) ((float) (1.0 - (1.0 - attack_specific_modifier) * p2));
+						}
 					}
 
 					Input var10000 = clientPlayer.input;
-					var10000.movementForward = (float) ((double) var10000.movementForward * multiplier);
+					var10000.movementForward = (float) ((double) var10000.movementForward * multiplier * attack_specific_modifier);
 					var10000 = clientPlayer.input;
-					var10000.movementSideways = (float) ((double) var10000.movementSideways * multiplier);
+					var10000.movementSideways = (float) ((double) var10000.movementSideways * multiplier * attack_specific_modifier);
 				}
 			}
 		}
-		var betterCombatExtensionServerConfig = BetterCombatExtension.SERVER_CONFIG;
 		boolean isMovementLockingDisabled = activeItemStack.isIn(BetterCombatExtension.DISABLES_MOVEMENT_LOCKING_DURING_ATTACK);
 		if (betterCombatExtensionServerConfig.enable_movement_locking_attacks.get() && !isMovementLockingDisabled && isWeaponSwingInProgress) {
 			boolean isVehicleDisablingMovementLocking = clientPlayer.getVehicle() != null && clientPlayer.getVehicle().getType().isIn(BetterCombatExtension.DISABLES_MOVEMENT_LOCKING_WHEN_RIDDEN);